+++ tests/postgres/table.rs
@@ -256,6 +256,34 @@ fn create_11() {
 fn create_12() {
     assert_eq!(
         Table::create()
+            .table(BinaryType::Table)
+            .col(ColumnDef::new(BinaryType::BinaryLen).binary_len(32))
+            .col(ColumnDef::new(BinaryType::Binary).binary())
+            .col(ColumnDef::new(BinaryType::BlobSize).blob(BlobSize::Blob(Some(32))))
+            .col(ColumnDef::new(BinaryType::TinyBlob).blob(BlobSize::Tiny))
+            .col(ColumnDef::new(BinaryType::Blob).blob(BlobSize::Blob(None)))
+            .col(ColumnDef::new(BinaryType::MediumBlob).blob(BlobSize::Medium))
+            .col(ColumnDef::new(BinaryType::LongBlob).blob(BlobSize::Long))
+            .to_string(PostgresQueryBuilder),
+        vec![
+            r#"CREATE TABLE "binary_type" ("#,
+            r#""binlen" bytea,"#,
+            r#""bin" bytea,"#,
+            r#""defb" bytea,"#,
+            r#""tb" bytea,"#,
+            r#""b" bytea,"#,
+            r#""mb" bytea,"#,
+            r#""lb" bytea"#,
+            r#")"#,
+        ]
+        .join(" ")
+    );
+}
+
+#[test]
+fn create_13() {
+    assert_eq!(
+        Table::create()
             .table(Char::Table)
             .col(ColumnDef::new(Char::Character).binary())
             .col(ColumnDef::new(Char::FontSize).binary_len(10))